pull request new friend form into its own component and scss file
