+++ client/src/scss/forms/_requestFriendForm.scss
@@ -1,0 +1,14 @@
+.addNewFriend {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  input {
+    border-radius: 10px;
+  }
+  ul {
+    .error {
+      color: red;
+      list-style: none;
+    }
+  }
+}
