+++ client/src/scss/index.scss
@@ -13,6 +13,7 @@
 // Forms styling
 @import "./forms/createAccountForm";
 @import "./forms/newMessageForm";
+@import "./forms/requestFriendForm";
 
 @import url("https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;500;700&family=Roboto+Mono:wght@300;400;700&display=swap");
