+++ client/src/scss/subComponents/_addFriends.scss
@@ -45,20 +45,7 @@
         }
       }
     }
-    .addNewFriend {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      input {
-        border-radius: 10px;
-      }
-      ul {
-        .error {
-          color: red;
-          list-style: none;
-        }
-      }
-    }
+
     .myRequests {
     }
   }
